test(runtime): harden timing and refactor oversized suites
Context:
- Several test modules had timing-sensitive waits, duplicated case matrices, and oversized files that were hard to review and maintain.
- We also needed canonical docs/status surfaces to reflect completion of the test-maintenance debt item.

Changes:
- replaced brittle sleep-based synchronization with event/barrier-driven waiting in scheduler/session lane tests and aligned related backstop assertions.
- split large test modules into focused suites (CLI, command surface, runtime conversation) and extracted shared test helpers.
- parameterized repetitive matrices in global config, security-language policy, and tool-dispatch tests.
- tightened test quality settings by lowering drill-sergeant max file length to 900 and enabling parallel test execution via xdist.
- synced status surfaces by closing the P3 oversized-test-module debt item with closure evidence.

Why:
- deterministic synchronization removes flake risk without changing production behavior.
- smaller, behavior-sliced suites plus matrix parameterization reduce maintenance overhead and improve reviewability.
- stricter file-length enforcement prevents future test-module bloat.

Validation:
- just lint (pass)
- just format-check (pass)
- just types (pass)
- just docs-check (pass)
- just test (pass, 311 passed)
- just quality && just test (pass)
- just status (pass)

Refs: docs/dev/debt/issues/debt-p3-split-oversized-test-modules.md
