+++ pyproject.toml
@@ -43,6 +43,7 @@ dev = [
     "docstr-coverage>=2.3",
     "pre-commit>=4.0",
     "pytest-drill-sergeant>=0.4.0",
+    "pytest-xdist>=3.8.0",
 ]
 
 [build-system]
@@ -64,7 +65,7 @@ drill_sergeant_enforce_aaa = true
 drill_sergeant_aaa_mode = "basic"
 drill_sergeant_aaa_severity = "error"
 drill_sergeant_enforce_file_length = true
-drill_sergeant_max_file_length = 1200
+drill_sergeant_max_file_length = 900
 drill_sergeant_file_length_mode = "error"
 markers = [
     "unit: Unit tests that test individual components in isolation",
